File uploading
I deleted the emoji button as there seemed to be little use but I can add it back later. I instead added two buttons, one to attach the file and the other to upload the file. The file uploading works, I just need to display the file name and have it appear in the chat.